refactor: Logout 위치 변경 및 MappingUrl 변경
